Add files via upload
This program saves your login credentials needed for different websites.
The program can also generate a random password for you if you want. 
And once that random password is generated, it is automatically copied to clipboard for you to paste into the website form.
The website, email/username, and password info is saved in a text file.