Add files via upload
This program is supposed to be a flash card app game to help learn Polish language, but it could be easily 
modified to learn any other subject.
If you know a word or the item on the front of the card, hit the checkmark, then it will show the answer on the 
backside. That word then gets sent to a file to keep track of words that you already know, so it won't show
you them repeatedly. If you don't know the word, hit the red button (X) and it will show what the answer is, and 
it will keep that word in the stash to show you again later on.
The timer could be turned on so the card turns over in 3 seconds. It is disabled for now. 
*Some performance improvements could be made in the future...